Task 7

See merge request orel-java-ee-training-2020/sergey_ashikhmin!7

+++ src/main/java/com/senla/courses/autoservice/service/GarageService.java
@@ -8,9 +8,9 @@ import com.senla.courses.autoservice.service.interfaces.IGarageService;
 import com.senla.courses.autoservice.service.interfaces.IMasterService;
 import com.senla.courses.autoservice.utils.ConsoleHelper;
 import com.senla.courses.autoservice.utils.CsvHelper;
+import com.senla.courses.autoservice.utils.SerializeUtil;
 
-import java.io.File;
-import java.io.FileNotFoundException;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,10 +18,14 @@ public class GarageService implements IGarageService {
 
     private IGarageDao garageDao;
     private IMasterService masterService;
+    private boolean addGaragePlaceOption = true;
+    private boolean removeGaragePlaceOption = true;
 
-    public GarageService(IGarageDao garageDAO, IMasterService masterService) {
+    public GarageService(IGarageDao garageDAO, IMasterService masterService, boolean addGaragePlaceOption, boolean removeGaragePlaceOption) {
         this.garageDao = garageDAO;
         this.masterService = masterService;
+        this.addGaragePlaceOption = addGaragePlaceOption;
+        this.removeGaragePlaceOption = removeGaragePlaceOption;
     }
 
     @Override
@@ -42,13 +46,23 @@ public class GarageService implements IGarageService {
 
     @Override
     public boolean addGaragePlace(int garageId, int garagePlaceId, String type, int area) {
-        GaragePlace garagePlace = new GaragePlace(garagePlaceId, garageId, type, area);
-        return garageDao.addGaragePlace(garagePlace);
+        if (addGaragePlaceOption) {
+            GaragePlace garagePlace = new GaragePlace(garagePlaceId, garageId, type, area);
+            return garageDao.addGaragePlace(garagePlace);
+        } else {
+            ConsoleHelper.writeMessage("Возможность добавления места в гараже отключена");
+            return false;
+        }
     }
 
     @Override
     public boolean removeGaragePlace(int garageId, int garagePlaceId) {
-        return garageDao.removeGaragePlace(findGarageById(garageId), findGaragePlaceById(garageId, garagePlaceId));
+        if (removeGaragePlaceOption) {
+            return garageDao.removeGaragePlace(findGarageById(garageId), findGaragePlaceById(garageId, garagePlaceId));
+        } else {
+            ConsoleHelper.writeMessage("Возможность удаления места в гараже отключена");
+            return false;
+        }
     }
 
     @Override
@@ -198,6 +212,16 @@ public class GarageService implements IGarageService {
         garagePlaceAsList.add(String.valueOf(garagePlace.getArea()));
         garagePlaceAsList.add(String.valueOf(garagePlace.isBusy()));
         return garagePlaceAsList;
+    }
+
+    @Override
+    public void saveState() {
+        SerializeUtil.saveState(getAllGarages(), "SerialsGarages.out");
+    }
+
+    @Override
+    public void loadState() {
+        garageDao.setAllGarages(SerializeUtil.loadState(Garage.class, "SerialsGarages.out"));
     }
 
 }
